fix(plasma): widen accumulator to uint16_t to avoid sum truncation

+++ AllEffects_FastLED/plasma.h
@@ -7,7 +7,7 @@ void plasma() {
   uint16_t t = millis() / 16;
   for (uint8_t y = 0; y < NUM_ROWS; y++) {
     for (uint8_t x = 0; x < NUM_COLS; x++) {
-      uint8_t v =
+      uint16_t v =
         sin8(x * 16 + t) +
         sin8(y * 32 + t * 2) +
         sin8((x + y) * 12 + t);
